Added support for generated tiles with no supports
This update adds support for tiles with no pregenerated supports. I implemented this in the first way I found possible, with out understanding the full code, so it might be a bit of a kludge, but it seems to work fine.

The openlock.blend file doesn't seem to be included in the repository in the same way it is in the release, so I only included the custom file I used to hold the supportless cutting meshes.